delete old code; move *many* functions; get rid of 'lexicalookups.py'
